Project 5: Computer
This is a final hardware component – computer as a whole, with ROM
instruction memory, RAM with Keyboard and Screen IO memory maps and a
CPU. It runs the program stored in the ROM chip starting from address 0
onward. It can be reset by sending a signal on the reset pin to restart
the program from the beginning.